+++ msr/raplcap-cpuid.c
@@ -84,6 +84,14 @@ int cpuid_is_cpu_supported(uint32_t family, uint32_t model) {
       //
       case CPUID_MODEL_CANNONLAKE_L:
       //
+      case CPUID_MODEL_ICELAKE:
+      case CPUID_MODEL_ICELAKE_L:
+      case CPUID_MODEL_ICELAKE_X:
+      case CPUID_MODEL_ICELAKE_D:
+      //
+      case CPUID_MODEL_COMETLAKE:
+      case CPUID_MODEL_COMETLAKE_L:
+      //
       case CPUID_MODEL_XEON_PHI_KNL:
       case CPUID_MODEL_XEON_PHI_KNM:
       //